Bug fix for clients that don't request the roster, but use presence.

+++ XMPPClient.h
@@ -27,7 +27,7 @@
 	NSString *password;
 	int priority;
 	
-	Byte flags;
+	UInt16 flags;
 	
 	XMPPStream *xmppStream;
 	NSError *streamError;
